+++ domain_details.go
@@ -1,0 +1,32 @@
+package godaddygo
+
+import (
+    "time"
+)
+
+// DomainDetails defines the details of a domain
+type DomainDetails struct {
+	AuthCode               string        `json:"authCode,omitempty"`
+	ContactAdmin           Contact       `json:"contactAdmin,omitempty"`
+	ContactBilling         Contact       `json:"contactBilling,omitempty"`
+	ContactRegistrant      Contact       `json:"contactRegistrant,omitempty"`
+	ContactTech            Contact       `json:"contactTech,omitempty"`
+	CreatedAt              time.Time     `json:"createdAt,omitempty"`
+	DeletedAt              time.Time     `json:"deletedAt,omitempty"`
+	TransferAwayEligibleAt time.Time     `json:"transferAwayEligibleAt,omitempty"`
+	Domain                 string        `json:"domain,omitempty"`
+	DomainID               int           `json:"domainId,omitempty"`
+	ExpirationProtected    bool          `json:"expirationProtected,omitempty"`
+	Expires                time.Time     `json:"expires,omitempty"`
+	ExposeWhois            bool          `json:"exposeWhois,omitempty"`
+	HoldRegistrar          bool          `json:"holdRegistrar,omitempty"`
+	Locked                 bool          `json:"locked,omitempty"`
+	NameServers            []string      `json:"nameServers,omitempty"`
+	Privacy                bool          `json:"privacy,omitempty"`
+	RenewAuto              bool          `json:"renewAuto,omitempty"`
+	RenewDeadline          time.Time     `json:"renewDeadline,omitempty"`
+	Status                 string        `json:"status,omitempty"`
+	SubAccountID           string        `json:"subAccountId,omitempty"`
+	TransferProtected      bool          `json:"transferProtected,omitempty"`
+	Verifications          Verifications `json:"verifications,omitempty"`
+}
