+++ domain_name.go
@@ -1,0 +1,6 @@
+package godaddygo
+
+// DomainName defines a domain name
+type DomainName struct {
+	Status string `json:"status,omitempty"`
+}
