+++ enum.go
@@ -1,0 +1,6 @@
+package godaddygo
+
+/*
+ * Enum-like vars
+ */
+
