+++ real_name.go
@@ -1,0 +1,6 @@
+package godaddygo
+
+// RealName defines the real name
+type RealName struct {
+	Status string `json:"status,omitempty"`
+}
